+++ src/utils/input_parser.cpp
@@ -24,7 +24,7 @@ inline Coordinates parse_coordinates(const rapidjson::Value& object,
       !object[key][0].IsNumber() or !object[key][1].IsNumber()) {
     throw InputException("Invalid " + std::string(key) + " array.");
   }
-  return {{object[key][0].GetDouble(), object[key][1].GetDouble()}};
+  return {object[key][0].GetDouble(), object[key][1].GetDouble()};
 }
 
 inline std::string get_string(const rapidjson::Value& object, const char* key) {