+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 - Use `std::chrono::milliseconds` for `Timeout` value (#728)
+- Use `struct` for storing `Coordinates` instead of an `std::array` (#730)
 
 ### Fixed